ENH: Add option to ignore any broken plugins.
Currently, if there are any broken plugins, the EggBasketPluginManager
just fails to return any plugins at all crashing an the application when
the application runs.  The added option which is off by default allows
the plugin manager to continue loading good plugins and returns them
while storing any errors for the application writer to act on.