REF : Refactor the version gen machinery in setup.py
Separate the tasks of getting the version info from generating the
version info. Further, separate the tasks of getting the version
info from the git source tree and from the version file.

Finally, dont import the install_requires from the package init file.

NOTE : The version file is not a private file anymore.

	modified:   .gitignore
	modified:   envisage/__init__.py
	modified:   setup.py