Refactor the version gen machinery in setup.py (#190)
* REF : Refactor the version gen machinery in setup.py

Separate the tasks of getting the version info from generating the
version info. Further, separate the tasks of getting the version
info from the git source tree and from the version file.

Finally, dont import the install_requires from the package init file.

NOTE : The version file is not a private file anymore.

	modified:   .gitignore
	modified:   envisage/__init__.py
	modified:   setup.py

* REF : Add the case for a released source archive

GitHub allows users to download the source archive as a tar/targz
file for a released version. The changes in this commit support that
usecase.

The error message when resolve_version fails has been slightly updated

	modified:   setup.py

* FIX : Correct the formatted string

	modified:   setup.py

* TST : Borrow tests for the version module

from traits

	new file:   envisage/tests/test_version.py

* FEAT : Use Git magic to add commit hash to archive

if a source archive is created by downloading a zip or tar archive
of the codebase, the ARCHIVE_COMMIT_HASH var in setup.py will get
updated to the commit hash of the HEAD commit when the archive
was generated

	new file:   .gitattributes
	modified:   setup.py

* REF : Check for a source archive first

and then handle the case when its a source distribution

	modified:   setup.py