Add a test to demonstrate current behaviour
Listening to "x_items" in the on_trait_change mini-language
means listening to mutation of a traits list/dict. However the
"x_items" here is actually just another trait, and the extension
point value is in fact a ephemeral property that gets
recomputed every time. Attempt to mutate it does nothing meaningful,
which I believe is intended, otherwise the extension point value
and the registry internal state will become inconsistent.