Remove reference back to the object
The reference only serves to support some edge cases where
the extension point values become out-of-sync because the listeners
are not connected yet. Once the listeners are connected the values
are refreshed. So all in all it is not worth keeping the reference,
as one will then need to use weakref to avoid cycle references,
and special handling in the pickling logic to handle that weakref.