Refresh layout pickles (#472)
This PR refreshes the two layout pickle files, by simply reading and then writing them again with Traits 6.3.2. This has the effect of changing the module used for importing TraitListObject from traits.trait_handlers to traits.trait_list_object.

I've also added a script that can be used to create the pickles directly if we need to do this again in the future.

Should fix #471.