Don't require ETS_TOOLKIT in tests (#526)
[PR includes commits from, and extends, #513]

This PR updates our testing machinery in two ways:

- We no longer rely on `ETS_TOOLKIT` (or `QT_API`) being set in order to
run tests that require a GUI; we just rely on Pyface's default
mechanisms for toolkit detection. (As a consequence, we retain the
option to _use_ `ETS_TOOLKIT` if it's helpful.)
- `etstool.py` support for backends other than those available through
EDM has been removed. Meanwhile, support for `pyqt6` has been added.
(Tests currently fail with PyQt6, due to a known issue in TraitsUI:
https://github.com/enthought/traitsui/issues/1926)

Closes #525
Closes #370