Fix TasksApplication lifecycle: don't stop plugins while event loop is running (#524)
This PR fixes a long-standing bug in the `TasksApplication` where the
plugins were stopped while the event loop was still running; that seems
unlikely to be what any particular application wants, and we've had to
work around it in downstream projects. I doubt that this was ever the
intended behaviour.

For backwards compatibility, we keep the behaviour where closing the
last Tasks windows causes the application to exit. (That behaviour is
also worth discussing, but likely in a separate issue.)

Closes #221.