Build updates and modernization (#513)
This PR updates the build machinery to use pyproject.toml, and drops
Python 3.6 support.

## Detailed changes

- Drop Python 3.6 support; we now require Python >= 3.7.
- Drop Python 3.6 from test workflows.
- Only use PySide6 for testing. (The main package still permits PySide2
and PyQt5 for Python versions for which they're available; we just don't
test on those.)
- Rename `setup.cfg` (which only contained `flake8` configuration
settings) to `.flake8`
- Replace `setup.py` with `pyproject.toml`
- Remove various outdated references to `setup.py`
- Remove `envisage.version` and `envisage.__version__` (and tests)
- Optional dependencies (a.k.a. `extras_require`) have been removed
- We no longer rely on `ETS_TOOLKIT` (or `QT_API`) being set in order to
run tests that require a GUI; we just rely on Pyface's default
mechanisms for toolkit detection. (As a consequence, we retain the
option to _use_ `ETS_TOOLKIT` if it's helpful.)
- `etstool.py` support for backends other than those available through
EDM has been removed. Meanwhile, support for `pyqt6` has been added.
(Tests currently fail with PyQt6, due to a known issue in TraitsUI:
https://github.com/enthought/traitsui/issues/1926)

Closes #370
Closes #511
Closes #512
Closes #519
Closes #525