Improve doc build; fix Read the Docs config (#560)
Our Read the Docs build is currently failing. The proximate cause is
that I removed the `docs` optional requirements when things got migrated
to `pyproject.toml`, but I suspect that even with that fixed other
things would have gone wrong.

This PR reworks the doc build a bit to make it more Read-the-Docs
friendly:

- Hook up the apidoc build in the Sphinx configuration file, so that we
only need the usual `sphinx-build -b html` step to build the docs; the
api documentation generation is taken care of for us. This ensures that
the API docs will be included in the Read the Docs build.
- Move the documentation requirements to `docs/requirements.txt`, and
pin versions. (Without this, RtD will use Sphinx 1.8.6. See
https://docs.readthedocs.io/en/stable/build-default-versions.html#external-dependencies
for more info.)
- Update the RtD configuration file.
- Update the doc test build workflow and `etstool.py` for the above
changes.