Automate documentation building (#565)
This PR automates documentation building for the Envisage repository. It
repurposes the current structure of the gh-pages branch, and adds
workflows for updating documentation on every push to main and every
release.

## New gh-pages structure

The new gh-pages structure is as follows:

- The top-level directory contains documentation that matches the
current in-development version of the codebase.
- The `<major>.<minor>` numbered subdirectories (`5.2`, `6.0`, `6.1`,
etc.) contain documentation for _released_ versions of Envisage

Note: as before, documentation for a _bugfix_ release (e.g., Envisage
7.1.1) overwrites the documentation for the previous release (e.g.,
Envisage 7.1.0). This is deliberate - documentation bugs may be fixed in
a bugfix release, but it should be rare for an API to change in a bugfix
release, and for any given `<major>.<minor>` version number users should
usually be seeing the latest documentation.

## Workflows

This PR adds two workflows for updating the docs:

- The `update-gh-pages.yml` workflow updates the in-development docs on
every push to the `main` branch (i.e., every PR merge)
- The `update-gh-pages-on-release.yml` workflow updates the release docs
on every published release.

Both workflows use the new `docs/update_gh_pages.py` script to do the work of
copying the newly-built docs to the appropriate place in the `gh-pages`
branch.

The workflows are very similar: the key differences between the two
workflows are the triggers, and the passing of the `--tag` argument to
`update_gh_pages.py` in the release case. Otherwise, the workflows are
essentially identical.

## Notes

Sufficient branch protections have been removed from the gh-pages branch to allow the workflows to push to that branch. The branch protections should still prevent force pushes and deletion, though.