Clean up and fix egg-based tests (#564)
This PR fixes mutable-global-state-related issues with the egg-based
tests (again).

The root cause of the problems is that we're interfacing with an
ill-defined (and now
[deprecated](https://github.com/pypa/setuptools/pull/3843) and
unmaintained) API in `setuptools` code, which makes those interactions
fragile, and that tests make significant changes to global state.

The immediate cause of the issues is that we were trying to undo changes
to `pkg_resources` global state by replacing the global
`pkg_resources.working_set` with a fresh `WorkingSet` object during test
teardown, but that `WorkingSet` object was missing some of the
initialization (registered callbacks) that the original `WorkingSet`
had.

The main change in this PR is to try to undo global state changes by
_mutating_ the existing `pkg_resources.working_set` back to its original
state, instead of replacing `pkg_resources.working_set`. Alongside that,
we've reorganised the relevant tests to share common machinery (now in
`envisage.tests.support`) and to make use of context managers for
cleanup.

Fixes #440 
Fixes #563 
The test suite also now runs cleanly under the `haas` test runner.